+++ src/app/common/io/base-application.ts
@@ -2,7 +2,6 @@ import { BrowserWindow } from 'electron';
 import { Observable } from 'rxjs';
 import { WindowSize } from './window-size';
 export abstract class BaseApplication {
-    public abstract argumentsReceived$: Observable<string[]>;
     public abstract getGlobal(name: string): unknown;
     public abstract getCurrentWindow(): BrowserWindow;
     public abstract getWindowSize(): WindowSize;
