--- src/app/services/drag-and-drop/base-drag-and-drop.service.ts
@@ -1,3 +1,0 @@
-export abstract class BaseDragAndDropService {
-    public abstract listenToOperatingSystemFileDrops(): void;
-}
