+++ src/app/services/event-listener/base-event-listener.service.ts
@@ -1,0 +1,7 @@
+import { Observable } from 'rxjs';
+
+export abstract class BaseEventListenerService {
+    public abstract argumentsReceived$: Observable<string[]>;
+    public abstract filesDropped$: Observable<string[]>;
+    public abstract listenToEvents(): void;
+}
