+++ src/app/services/file/base-file.service.ts
@@ -1,5 +1,4 @@
 export abstract class BaseFileService {
     public abstract hasPlayableFilesAsParameters(): boolean;
     public abstract enqueueParameterFilesAsync(): Promise<void>;
-    public abstract enqueueGivenParameterFilesAsync(parameters: string[]): Promise<void>;
 }
